#71 Move SQLite detection and path-finding logic to sqlitedetection.go
