+++ internal/sqlite/sqlite.go
@@ -15,9 +15,7 @@ import (
 	"fmt"
 	"io"
 	"log/slog"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"regexp"
 	"runtime"
 	"strings"
@@ -105,131 +103,6 @@ func (e *Engine) GetVersion() (string, error) {
 		return "", err
 	}
 	return strings.TrimSpace(string(output)), nil
-}
-
-// GetPath returns the full path to the SQLite binary
-func (e *Engine) GetPath() (string, error) {
-	return exec.LookPath(e.Bin)
-}
-
-// getLinuxAptSQLitePaths returns common apt SQLite installation paths on Linux
-func getLinuxAptSQLitePaths() []string {
-	if runtime.GOOS != "linux" {
-		return nil
-	}
-
-	// Common locations where apt installs sqlite3
-	return []string{
-		"/usr/bin/sqlite3",
-		"/usr/local/bin/sqlite3",
-		"/bin/sqlite3",
-		"/usr/sbin/sqlite3",
-	}
-}
-
-// getWinGetSQLitePaths returns common WinGet SQLite installation paths on Windows
-func getWinGetSQLitePaths() []string {
-	if runtime.GOOS != "windows" {
-		return nil
-	}
-
-	paths := []string{}
-
-	// Common SQLite package directory patterns
-	sqlitePatterns := []string{
-		"SQLite.SQLite_Microsoft.Winget.Source_*",
-		"SQLite.SQLite_*",
-	}
-
-	// 1. User-level installation (non-elevated)
-	userProfile := os.Getenv("USERPROFILE")
-	if userProfile != "" {
-		userWinGetPath := filepath.Join(userProfile, "AppData", "Local", "Microsoft", "WinGet", "Packages")
-
-		for _, pattern := range sqlitePatterns {
-			fullPattern := filepath.Join(userWinGetPath, pattern)
-			matches, err := filepath.Glob(fullPattern)
-			if err == nil {
-				for _, match := range matches {
-					paths = append(paths, filepath.Join(match, "sqlite3.exe"))
-				}
-			}
-		}
-	}
-
-	// 2. System-level installation (elevated/admin)
-	programFiles := os.Getenv("ProgramFiles")
-	if programFiles != "" {
-		systemWinGetPath := filepath.Join(programFiles, "WinGet", "Packages")
-
-		for _, pattern := range sqlitePatterns {
-			fullPattern := filepath.Join(systemWinGetPath, pattern)
-			matches, err := filepath.Glob(fullPattern)
-			if err == nil {
-				for _, match := range matches {
-					paths = append(paths, filepath.Join(match, "sqlite3.exe"))
-				}
-			}
-		}
-	}
-
-	// 3. Alternative system location (some versions use this)
-	programData := os.Getenv("ProgramData")
-	if programData != "" {
-		altSystemWinGetPath := filepath.Join(programData, "Microsoft", "WinGet", "Packages")
-
-		for _, pattern := range sqlitePatterns {
-			fullPattern := filepath.Join(altSystemWinGetPath, pattern)
-			matches, err := filepath.Glob(fullPattern)
-			if err == nil {
-				for _, match := range matches {
-					paths = append(paths, filepath.Join(match, "sqlite3.exe"))
-				}
-			}
-		}
-	}
-
-	return paths
-}
-
-// findSQLiteInApt searches for SQLite in apt installation directories
-func (e *Engine) findSQLiteInApt() (string, error) {
-	if runtime.GOOS != "linux" {
-		return "", fmt.Errorf("apt search only available on Linux")
-	}
-
-	paths := getLinuxAptSQLitePaths()
-	for _, path := range paths {
-		if _, err := os.Stat(path); err == nil {
-			// Test if the executable works
-			cmd := exec.Command(path, "-version")
-			if err := cmd.Run(); err == nil {
-				return path, nil
-			}
-		}
-	}
-
-	return "", fmt.Errorf("SQLite not found in standard apt installation directories")
-}
-
-// findSQLiteInWinGet searches for SQLite in WinGet installation directories
-func (e *Engine) findSQLiteInWinGet() (string, error) {
-	if runtime.GOOS != "windows" {
-		return "", fmt.Errorf("WinGet search only available on Windows")
-	}
-
-	paths := getWinGetSQLitePaths()
-	for _, path := range paths {
-		if _, err := os.Stat(path); err == nil {
-			// Test if the executable works
-			cmd := exec.Command(path, "-version")
-			if err := cmd.Run(); err == nil {
-				return path, nil
-			}
-		}
-	}
-
-	return "", fmt.Errorf("SQLite not found in WinGet installation directories")
 }
 
 // GetPathWithPackageManager returns the full path to the SQLite binary, checking package manager locations
